// src/modules/category/enums/category-message.enum.ts
export enum CategoryMessage {
  CREATED = 'Category successfully created.',
  RETRIEVED_ALL = 'List of all categories retrieved successfully.',
  RETRIEVED_ONE = 'Category retrieved successfully.',
  UPDATED = 'Category updated successfully.',
  DELETED = 'Category deleted successfully.',
  NOT_FOUND = 'Category not found.',
}